Move build-url into separate lib
